refactor: rename manager module to instance_manager and update references
